add test for extencion and the proyect structure is improved and the customization option is added
