+++ package.json
@@ -2,7 +2,7 @@
   "name": "drawfolderstructure",
   "displayName": "Draw Folder Structure",
   "description": "",
-  "version": "0.1.5",
+  "version": "1.0.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/krivoox/draw-folder-structure"
