+++ src/functions/should-exclude.ts
@@ -1,0 +1,3 @@
+export function shouldExclude(name: string, patterns: string[]): boolean {
+  return patterns.some((pattern) => name.includes(pattern));
+}
