+++ src/test/suite/extension.test.ts
@@ -1,15 +1,29 @@
 import * as assert from 'assert';
-
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
-import * as vscode from 'vscode';
-// import * as myExtension from '../../extension';
+import * as path from 'path';
+import { generateStructure } from '../../functions/generate-structure';
+import { shouldExclude } from '../../functions/should-exclude';
 
 suite('Extension Test Suite', () => {
-	vscode.window.showInformationMessage('Start all tests.');
+  test('generateStructure should return correct structure', () => {
+    const testFolderPath = path.resolve(
+      __dirname,
+      '../../../src/test/suite/testFolder'
+    );
+    const exclude = ['.git'];
+    const structure = generateStructure(testFolderPath, 0, exclude);
+    const expectedStructure = `├── README.md
+└── src
+  ├── app.ts
+  └── index.ts
+`;
 
-	test('Sample test', () => {
-		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
-		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
-	});
+    assert.strictEqual(structure, expectedStructure);
+  });
+
+  test('shouldExclude should return true for excluded patterns', () => {
+    const patterns = ['.git', 'node_modules'];
+    assert.strictEqual(shouldExclude('.git', patterns), true);
+    assert.strictEqual(shouldExclude('node_modules', patterns), true);
+    assert.strictEqual(shouldExclude('src', patterns), false);
+  });
 });
